+++ magda-int-test/src/test/scala/au/csiro/data61/magda/api/RegionsApiSpec.scala
@@ -255,6 +255,21 @@ class RegionsApiSpec
     }
   }
 
+  it(
+    "should only return Test region record regionId=10101 regionType=SA3 when specify regionId=10101 type=SA3"
+  ) {
+    Get(s"/v0/regions?type=SA3&regionId=10101") ~> addTenantIdHeader ~> api.routes ~> check {
+      status shouldBe OK
+      contentType shouldBe `application/json`
+      val response = responseAs[RegionSearchResult]
+
+      response.hitCount shouldEqual 1
+
+      response.regions.head.queryRegion.regionType shouldEqual "SA3"
+      response.regions.head.queryRegion.regionId shouldEqual "10101"
+    }
+  }
+
   val indexedRegions = List(
     (
       new RegionSource(