+++ CHANGES.md
@@ -48,6 +48,7 @@ UI:
 -   Disable edit button on harvested datasets
 -   Added the support of processing ESRI REST API URL
 -   Use IBM Plex Sans font
+-   Added `add dataset review page`
 
 Storage:
 
@@ -86,6 +87,7 @@ Authorization:
 Search
 
 -   Began work on a NodeJS-based search API to replace the scala one
+-   Allow fetch region record by region id
 
 Others: